LocWeather.ipynb
Created ipynb to check N entries of the loc table for weather data that
is less than 4 hours and less than 100 miles from the measured_at and
lat lon of the entry. It then stores the data in the loc_weather table
and updates the locs table column weatherchecked to 1.