6-21 Update
Added Database QN.ipynb which displays the results of a query number
from SQL.txt. Updated InvalidPoints-WadeCopy2.ipynb to show regression
plots. It also displays the predictions on a map, and clicking a point
will display a popup if the point was incorrectly marked by Random
Forest. False Positives are displayed as orange points, while False
Negatives are displayed as green. True Positives are blue, and True
Negatives are red. Updated Classification-WadeCopy.py to return the
predicted values of each technique.