Created and moved files
Created PredictingPoints.ipynb that uses the trained classification
algorithm to predict points.
Created trainedclassification.p which is a saved python object of the
trained classification algo. It was made from pickle.
Created numtimesinvalid which is a python object created using pickle
that stores the number of times lat lon pairs are marked invalid.
Moved many old files to a newly created OLD folder.